corrected some mistakes
Transferred data to the right folders, corrected the pattern
