--- src/components/Home/Home.module.scss
@@ -1,119 +1,0 @@
-@use '../../styles/mixins';
-
-.form {
-    margin-top: 32px;
-    width: 320px;
-    height: 568px;
-    @include mixins.mq(tablet) {
-        margin-top: 100px;
-        width: 768px;
-        height: 944px;
-    }
-    @include mixins.mq(desktop) {
-        margin-top: 138px;
-        width: 608px;
-        height: 345px;
-    }
-}
-.formParts {
-    margin-bottom: 40px;
-    display: flex;
-    flex-direction: column;
-    @include mixins.mq(tablet) {
-        justify-content: left;
-        flex-direction: row;
-        margin-bottom: 60px;
-    }
-    @include mixins.mq(desktop) {
-        justify-content: space-between;
-    } 
-}
-.formPart {
-    width: 240px;
-    @include mixins.mq(tablet) {
-        margin-right: 32px;
-    }
-}
-.label {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.title {
-    margin-bottom: 32px;
-    font-weight: 700;
-    font-size: 18px;
-    line-height: 1.444;
-    color: #212121; 
-    @include mixins.mq(tablet) {
-        margin-bottom: 68px;
-        width: 640px;
-        height: 85px;
-        font-size: 34px;
-        line-height: 1.205;
-    }
-    @include mixins.mq(desktop) {
-        width: 608px;
-        height: 85.5px;
-    }   
-}
-.checkbox {
-    display: none;
-}
-.fake {
-    display: inline-block;
-    width: 20px;
-    height: 20px;
-    border: 1px solid #E0E0E0;
-    position: relative;
-    border-radius: 50%;
-}
-.fake::before {
-    content: "";
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    display: block;
-    width: 10px;
-    height: 10px;
-    background-color: var(--accent-color);
-    opacity: 0;
-    transition: .2s;
-    border-radius: 50%;
-}
-.text {
-    padding-left: 7px;
-    color: var(--second-text-color);
-    font-weight: 700;
-    font-size: 14px;
-    line-height: 1.649;
-    letter-spacing: 0.04em;
-}
-.checkbox:checked + .fake::before {
-    opacity: 1;
-}
-.checkbox:checked ~ .text {
-    color: var(--accent-color);
-}
-.radioBlock {
-    margin-top: -30px;
-    display: flex;
-    flex-direction: row;
-}
-.listRadio {
-    margin-right: 30px;
-    display:flex;
-    align-items: center;
-    justify-content: center;
-}
-.buttonPosition {
-    display: flex;
-    justify-content: center;
-    @include mixins.mq(tablet) {
-    justify-content: left;
-    }
-    @include mixins.mq(desktop) {
-        margin-left: 323px;
-    }
-}
