+++ src/components/Shared/TextField/fields.jsx
@@ -32,33 +32,37 @@ export const field = {
     type: 'text',
     placeholder: 'Height',
     required: true,
-    //pattern and title
+    pattern: '[0-9]{2,3}',
+    title: 'Please enter a valid number between 0 and 250.',
   },
   age: {
     name: 'age',
     type: 'text',
     placeholder: 'Age',
     required: true,
-    pattern: '[0-1]{1}[0-9]{0,2}',
-    title: 'Please enter a valid number between 0 and 200.',
+    pattern: '[0-9]{2,3}',
+    title: 'Please enter a valid number between 0 and 100.',
   },
   weight: {
     name: 'weight',
     type: 'text',
     placeholder: 'Current weight',
     required: true,
+    pattern: '[0-9]{2,3}',
+    title: 'Please enter a valid number between 0 and 200.',
   },
   desiredWeight: {
     name: 'desiredWeight',
     type: 'text',
     placeholder: 'Desired weight',
     required: true,
+    pattern: '[0-9]{2,3}',
+    title: 'Please enter a valid number between 0 and 200.',
   },
   bloodType: {
     name: 'blodtype',
     type: '',
     placeholder: 'Blood Type',
-    pattern: '([AaBbOo]|[Aa][Bb])[+-]',
     title: '',
   },
 };
