+++ src/redux/store.js
@@ -11,6 +11,7 @@ import {
 } from 'redux-persist';
 import storage from 'redux-persist/lib/storage';
 
+import dailyRateReducer from './daily-rate/daily-rate-slice'
 import auth from './auth/auth-slice';
 
 const persistConfig = {
@@ -24,6 +25,7 @@ const persistedReducer = persistReducer(persistConfig, auth);
 export const store = configureStore({
   reducer: {
     auth: persistedReducer,
+    dailyRate: dailyRateReducer,
   },
   middleware: getDefaultMiddleware =>
     getDefaultMiddleware({
